This commit adds server-side encryption using keys derived from a keyczar HMAC Signer; key generation is determisitic with respect to the HMAC and its key, so the server always rederives the same keys, hence can read its files.
On encryption, the server writes a metadata file containing the IV, the hmac,
the name, and the size, and it HMACs this data for decryption time.

The current implementation here uses AES-CBC and HMAC-SHA1 and matches the
original CloudProxy in doing streaming encryption instead of breaking up
encrypted files into blocks.

+++ Code/cloudproxy/file_server.cc
@@ -12,6 +12,7 @@ using std::mutex;
 namespace cloudproxy {
 
 FileServer::FileServer(const string &file_path,
+           const string &meta_path,
 	       const string &tls_cert,
 	       const string &tls_key,
 		const string &tls_password,
@@ -27,16 +28,35 @@ FileServer::FileServer(const string &file_path,
       public_policy_keyczar,
       public_policy_pem,
       acl_location,
-      server_key_location,
       host,
       port),
-    file_path_(file_path) {
-  // check to see if this path actually exists
+    main_key_(keyczar::Signer::Read(server_key_location.c_str())),
+    enc_key_(new string()),
+    hmac_key_(new string()),
+    file_path_(file_path),
+    meta_path_(meta_path) {
+  LOG(INFO) << "now in the file server constructor";
+  // check to see if these paths actually exist
   struct stat st;
   CHECK_EQ(stat(file_path_.c_str(), &st), 0) << "Could not stat the directory "
     << file_path_;
-
   CHECK(S_ISDIR(st.st_mode)) << "The path " << file_path_  << " is not a directory";
+
+  CHECK_EQ(stat(meta_path_.c_str(), &st), 0) << "Could not stat the directory "
+    << meta_path_;
+  CHECK(S_ISDIR(st.st_mode)) << "The path " << meta_path_  << " is not a directory";
+
+  // get binary data from the hmac
+  main_key_->set_encoding(keyczar::Keyczar::NO_ENCODING);
+
+  LOG(INFO) << "About to derive keys";
+
+  // generate keys
+  CHECK(DeriveKeys(main_key_.get(), &enc_key_, &hmac_key_)) << "Could not derive"
+    " keys for authenticated encryption";
+
+  CHECK(DeriveKeys(main_key_.get(), &enc_key_, &hmac_key_)) << "Could not"
+    " derive enc and hmac keys for authenticated encryption";
 }
 
 bool FileServer::HandleCreate(const Action &action, BIO *bio, string *reason,
@@ -51,6 +71,7 @@ bool FileServer::HandleCreate(const Action &action, BIO *bio, string *reason,
   // TODO(tmroeder): make this locking more fine-grained so that locks only
   // apply to individual files. Need a locking data structure for this.
   string path = file_path_ + string("/") + action.object();
+  string meta_path = meta_path_ + string("/") + action.object();
   {
     lock_guard<mutex> l(data_m_);
     struct stat st;
@@ -60,17 +81,29 @@ bool FileServer::HandleCreate(const Action &action, BIO *bio, string *reason,
       return false;
     }
 
-    FILE *f = fopen(path.c_str(), "w");
-    if (nullptr == f) {
+    if (stat(meta_path.c_str(), &st) == 0) {
+      LOG(ERROR) << "File " << meta_path << " already exists";
+      reason->assign("Already exists");
+      return false;
+    }
+
+    ScopedFile f(fopen(path.c_str(), "w"));
+    if (nullptr == f.get()) {
       LOG(ERROR) << "Could not create the file " << path;
       reason->assign("Could not create the file");
       return false;
     }
 
-    fclose(f);
+    ScopedFile mf(fopen(meta_path.c_str(), "w"));
+    if (nullptr == mf.get()) {
+      LOG(ERROR) << "Could not create the file " << meta_path;
+      reason->assign("Could not create the file");
+      return false;
+    }
   }
   
-  LOG(INFO) << "Create the file " << path;
+  LOG(INFO) << "Created the file " << path << " and its metadata " <<
+    meta_path;
   return true;
 }
 
@@ -83,6 +116,7 @@ bool FileServer::HandleDestroy(const Action &action, BIO *bio, string *reason,
   }
 
   string path = file_path_ + string("/") + action.object();
+  string meta_path = meta_path_ + string("/") + action.object();
   {
     lock_guard<mutex> l(data_m_);
     struct stat st;
@@ -92,8 +126,23 @@ bool FileServer::HandleDestroy(const Action &action, BIO *bio, string *reason,
       return false;
     }
 
+    if (stat(meta_path.c_str(), &st) != 0) {
+      LOG(ERROR) << "File " << meta_path << " does not exist";
+      reason->assign("Does not exist");
+      return false;
+    }
+
+    // ideally, this should be transactional, since the current instantiation of
+    // this code can get files into a state that can't be written and can't be
+    // destroyed
     if (unlink(path.c_str()) != 0) {
       LOG(ERROR) << "Could not unlink the file " << path;
+      reason->assign("Could not delete the file");
+      return false;
+    }
+
+    if (unlink(meta_path.c_str()) != 0) {
+      LOG(ERROR) << "Could not unlink the file " << meta_path;
       reason->assign("Could not delete the file");
       return false;
     }
@@ -104,14 +153,19 @@ bool FileServer::HandleDestroy(const Action &action, BIO *bio, string *reason,
 
 bool FileServer::HandleWrite(const Action &action, BIO *bio, string *reason,
 		    bool *reply, CloudServerThreadData &cstd) {
-
-
   string path = file_path_ + string("/") + action.object();
+  string meta_path = meta_path_ + string("/") + action.object();
   {
     lock_guard<mutex> l(data_m_);
     struct stat st;
     if (stat(path.c_str(), &st) != 0) {
       LOG(ERROR) << "File " << path << " does not exist";
+      reason->assign("Does not exist");
+      return false;
+    }
+
+    if (stat(meta_path.c_str(), &st) != 0) {
+      LOG(ERROR) << "File " << meta_path << " does not exist";
       reason->assign("Does not exist");
       return false;
     }
@@ -127,8 +181,10 @@ bool FileServer::HandleWrite(const Action &action, BIO *bio, string *reason,
       return false;
     }
   
-    if (!ReceiveStreamData(bio, path)) {
-      LOG(ERROR) << "Could not receive data from the client";
+    if (!ReceiveAndEncryptStreamData(bio, path, meta_path, action.object(),
+          enc_key_, hmac_key_, main_key_.get())) {
+      LOG(ERROR) << "Could not receive data from the client and write it"
+        " encrypted to disk";
       reason->assign("Receiving failed");
       return false;
     }
@@ -140,11 +196,18 @@ bool FileServer::HandleWrite(const Action &action, BIO *bio, string *reason,
 bool FileServer::HandleRead(const Action &action, BIO *bio, string *reason,
 		    bool *reply, CloudServerThreadData &cstd) {
   string path = file_path_ + string ("/") + action.object();
+  string meta_path = meta_path_ + string("/") + action.object();
   {
     lock_guard<mutex> l(data_m_);
     struct stat st;
     if (stat(path.c_str(), &st) != 0) {
       LOG(ERROR) << "File " << path << " does not exist";
+      reason->assign("Does not exist");
+      return false;
+    }
+
+    if (stat(meta_path.c_str(), &st) != 0) {
+      LOG(ERROR) << "File " << meta_path << " does not exist";
       reason->assign("Does not exist");
       return false;
     }
@@ -160,7 +223,8 @@ bool FileServer::HandleRead(const Action &action, BIO *bio, string *reason,
       return false;
     }
   
-    if (!SendStreamData(path, st.st_size, bio)) {
+    if (!DecryptAndSendStreamData(path, meta_path, action.object(),
+          bio, enc_key_, hmac_key_, main_key_.get())) {
       LOG(ERROR) << "Could not stream data from the file to the client";
       reason->assign("Could not stream data to the client");
       return false;
